integrate anyhow
as your application grows here, I think you'll come to appreciate
anyhow. A common suggestion is "thiserr for libraries, anyhow for
applications," and so we're following that. You're effectively putting
main in the library, which isn't bad, but that's why this is not
*strictly* following the letter of that advice, but instead, the spirit.

.context will give you output that looks like this:

```
Error: current directory should exist

Caused by:
    No such file or directory (os error 2)
```

I also tweaked the error messages, given that it not existing is the
only issue. For example, maybe the problem is permissions, in which
case, the error would look like this:

```
Error: current directory should exist

Caused by:
    Permission deined (os error 13)
```

which is uh, wrong. So as of this commit, this will instead say

```
Error: problem fetching current directory

Caused by:
    Permission deined (os error 13)
```

or

```
Error: problem fetching current directory

Caused by:
    No such file or directory (os error 2)
```

The other two messages are already pretty good, I'd say.