Re-work errors slightly
Okay so this is, in my mind, slightly better error handling. Instead of
erasing everything to a String, you're actually holding onto the various
error types. This would allow your users to handle errors in a more
granular way if they wished.

I struggled a bit for naming here. Network IO is still IO, so these feel
a bit redundant. However, I suspect separating these two does make sense
in some senese, as what you'd do to fix them is different. And merging
them into one grand IO error is actually a huge hassle and not helpful
for users. So I kept it this way. It's up to taste, really.